Recreation of Tasha's figures
Appears to be an issue with the colour bar for some of the plots and issues with 'smoothing' for others. The later may be due to fivedegreebins = 0, rather than 1, but this is what was ran in the original notebook. So I'm not totally sure of the issue. Added comments to these cells. Edited all the file paths so that they point to the correct files now, most of these where removing /Daviddata/ as it seemed like an unnecessary directory. I couldn't find the exact original file Wuhu_IonTr_run_SpE_Output_Dec-Feb_90-150km.nc, so I opted to use Wuhu_IonTr_run_SpE_Output_Dec-Feb_90-150km_0.25sigma_2xMpza_1xpeak.nc.